fix another spacing and refactor into seperate components
